// frontend/src/modules/user-account/models/form.model.ts
export class Form {
    Id: number = 0;
    FormId: number = 0;
    NameForm: string = '';
    Description: string = '';
    Orden: number = 0;
    Route: string = '';
    Controller: string = '';
    IconCss: string = '';
    ShowMenu: boolean = true;
    IsFormRegister: boolean = false;
    ModuleId: number = 0;
}